requirements.txt
Pillow
pillow-heif
